rm acquirelock from rocksdb and move tests
